built basic chatbot that can cite documentation
built a very basic notebook that can access text in a seperate directory to answer its questions and is able to return the documents used to inform its answers.
this can quite easily be adapted for the business application of answering ed questions about models.

Further research could be done into a persistent vectorstore to avoid having to re-embed documetns